requirements.txt
fastapi==0.95.0
bcrypt==4.0.1
mysql-connector-python==8.0.33
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
SQLAlchemy==2.0.13
uvicorn==0.21.1
passlib==1.7.4
